feat: keep keys in compressed form (#6753)
Description
---
Keeps all crypto public keys in compressed form, and only uncompressing
when needed.

Motivation and Context
---
This should improve speed and reduce the compression of keys. Comparing
unit test speeds this PR is roughly the same as before. But unit test
almost always do crypto operations to ensure everything is correct.
Where this PR increases performance is when only reading something and
providing it. For example, just reading the mainnet gen block in its
current form on avg it takes 56 seconds. With this pr this time is taken
down to on avg 15 seconds. Thats a 370% speed improvement.

How Has This Been Tested?
---
unit tests