fix!: monero seed validation (#6850)
Description
---
Fixes monero seed validation to always check seeds. 

Motivation and Context
---
Block 26320 is bad block breaking consensus. It has a RandomX VM key of
91ef83186cefaa646dc4c6e950e68e4debab52b4f4a9b7f465891e91fe5f6ce4
this key was used between height 2729 and 843, which is about what you
would expect from the monero consensus but was reused only once on block
26320 with hash
999785e3bb6a43189c7236de7c9720df27d56f6b9724cd22f5115a097d3f770e.

This PR will bypass the hash checking so that we can accept this block
in the chain, but only this block. It also fixes the validation bug that
caused the RandomX Seeds to not be added the database for tracking.

How Has This Been Tested?
---
Manual sync

Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [X] Other - Please specify

Wil whitelist a bad block to be accepted into the chain, and clear out
all bad blocks listed


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Streamlined block processing and chain reorganization by removing
legacy consensus parameters.

- **New Features**
- Enhanced the handling and logging of specialized proof-of-work data in
block headers.
- Updated the migration process to support a new database schema version
and to clear legacy records.
- Introduced a mechanism to allow specific block headers to bypass
certain validation checks, ensuring smoother validation flow.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->