fix(comms/yamux): dont poll the substream after closing/error (#6911)
Description
---
fix(comms/yamux): dont poll the substream after closing/error

Motivation and Context
---
We inadvertently polled the connection after closing it/after an error.
This is not correct and results in a panic

Ref https://github.com/libp2p/rust-yamux/pull/199
The PR above prevents the panic and sets the connection state to closed.

This related PR can also be merged before/after this one:
https://github.com/tari-project/tari/pull/6909

How Has This Been Tested?
---
Not explicitly tested

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Improved connection handling to prevent further processing once a
connection is closed, enhancing reliability by ensuring that operations
halt gracefully when encountering a closed state.
- Added tracking for the closed state of the worker, enhancing
robustness in connection management.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: SW van Heerden <swvheerden@gmail.com>