feat!: set tonic msg limits (#7016)
Description
---
Set tonic client _and server_ msg limits to 8MiB

~~**Note:** To set the corresponding limits on the server a tonic
upgrade is required as that is only supported in `tonic = "0.13.0"` and
we use `tonic = "0.12.3"`.~~

Motivation and Context
---
Block templates need to be send via grpc and we need to allow at least
6MiB (2MiB for coinbases and 4MiB for the transaction payload), whereas
the default tonic implementation limits this to 4MiB.

How Has This Been Tested?
---
Not tested - System-level testing [**TBD**].

What process can a PR reviewer use to test or verify this change?
---
- Code review.
- System-level testing.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
BREAKING CHANGE: These limits are set for the base node, wallet, merge
mining proxy and miner, but not for _external-to-main-tari-project_
applications like sha_p2pool or glytex.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added configuration to set maximum gRPC message size (8 MiB) for
wallet and base node clients, allowing handling of larger messages.

- **Refactor**
- Simplified internal logic for transaction handling in database and
validation test code, improving code clarity and consistency.
  - Enhanced test error messages for clearer diagnostics.

- **Chores**
- Updated dependencies to include a shared constant for maximum gRPC
message size.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->