Feat: Windows Build Instructions Revision (#6734)
Description
---

Mildly revised the instructions for installing the various dependencies
required to build Tari on Windows, incorporating some additional
comments from others who attempted to run it on Windows 11. I've also
revised the main readme to link to these instructions in the Windows
section.

Motivation and Context
---
Lee updated the Windows build instructions to account for the latest
changes to the process. I've revised it slightly and removed some
extraneous info from some of the sample boxes.

How Has This Been Tested?
---
Not tested as I only have a Mac machine for work.

What process can a PR reviewer use to test or verify this change?
---
Run through the instructions on a clean Windows environment and confirm
that Tari can be successfully built from source.


<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Streamlined the README by replacing detailed Windows setup
instructions with a link to an external, dedicated setup guide.
- Enhanced the external Windows development environment guide with
clearer, detailed instructions, troubleshooting tips, and comprehensive
explanations for setting up and building Tari protocol tools on Windows.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->