feat: block seed words extraction from view only wallets (#7022)
Description
---

Block the extraction of seed words from wallets that don't have access
to the correct seed words, which is view only wallets and hardware
enabled wallets.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Bug Fixes**
- Exporting seed words is now restricted to supported wallet types only.
Users with Hardware or View-only wallets will no longer be able to
export seed words, and will receive an appropriate error message if they
attempt to do so.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->